fix(storage) reusable configuration row component
