+++ src/components/ConfigurationRow.tsx
@@ -1,26 +1,24 @@
 import React, { ReactElement, ReactNode } from "react";
-import { Button, Icon, Label, useNotify } from "@canonical/react-components";
-import { SharedFormikTypes } from "pages/instances/forms/sharedFormTypes";
-import { useQuery } from "@tanstack/react-query";
-import { queryKeys } from "util/queryKeys";
-import { fetchProfiles } from "api/profiles";
-import { useParams } from "react-router-dom";
+import { Button, Icon, Label } from "@canonical/react-components";
 import { CpuLimit, MemoryLimit } from "types/limits";
 import { MainTableRow } from "@canonical/react-components/dist/components/MainTable/MainTable";
-import { EditInstanceFormValues } from "pages/instances/EditInstanceForm";
-import { LxdProfile } from "types/profile";
-import { figureInheritedValue } from "util/instanceConfigInheritance";
 import classnames from "classnames";
+import { FormikProps } from "formik/dist/types";
 
 interface Props {
-  formik: SharedFormikTypes;
+  formik: FormikProps<unknown>;
   name: string;
   label: string | ReactNode;
   children: ReactElement;
   defaultValue?: string | CpuLimit | MemoryLimit;
   disabled?: boolean;
-  readOnlyRenderer?: (value: unknown) => string | ReactNode;
   help?: string;
+  isOverridden: boolean;
+  inheritedValue: string | ReactNode;
+  inheritSource: string;
+  isReadOnly: boolean;
+  value: string | undefined;
+  overrideValue: string | ReactNode;
 }
 
 export const getConfigurationRow = ({
@@ -30,27 +28,14 @@ export const getConfigurationRow = ({
   children,
   defaultValue,
   disabled = false,
-  readOnlyRenderer,
   help,
+  isOverridden,
+  inheritedValue,
+  inheritSource,
+  isReadOnly,
+  value,
+  overrideValue,
 }: Props): MainTableRow => {
-  const notify = useNotify();
-  const { project } = useParams<{ project: string }>();
-  let profiles: LxdProfile[] = [];
-  if (project) {
-    const { data = [], error: profileError } = useQuery({
-      queryKey: [queryKeys.profiles],
-      queryFn: () => fetchProfiles(project),
-    });
-    profiles = data;
-
-    if (profileError) {
-      notify.failure("Loading profiles failed", profileError);
-    }
-  }
-
-  const values = formik.values as unknown as Record<string, string | undefined>;
-  const isOverridden = values[name] !== undefined;
-
   const toggleDefault = () => {
     if (isOverridden) {
       formik.setFieldValue(name, undefined);
@@ -60,12 +45,6 @@ export const getConfigurationRow = ({
     }
   };
 
-  const [inheritedValue, inheritSource] = figureInheritedValue(
-    formik.values,
-    name,
-    profiles
-  );
-  const isReadOnly = (formik.values as EditInstanceFormValues).readOnly;
   const getForm = (): ReactNode => {
     return (
       <div className="override-form">
@@ -75,7 +54,7 @@ export const getConfigurationRow = ({
             name: name,
             onBlur: formik.handleBlur,
             onChange: formik.handleChange,
-            value: values[name],
+            value: value,
           })}
         </div>
         <div>
@@ -90,15 +69,6 @@ export const getConfigurationRow = ({
         </div>
       </div>
     );
-  };
-
-  const getInheritedValue = (): ReactNode => {
-    return readOnlyRenderer ? readOnlyRenderer(inheritedValue) : inheritedValue;
-  };
-
-  const getOverrideValue = (): ReactNode => {
-    const value = values[name] === "" ? "-" : values[name];
-    return readOnlyRenderer ? readOnlyRenderer(value) : value;
   };
 
   const displayLabel = isOverridden ? (
@@ -128,7 +98,7 @@ export const getConfigurationRow = ({
         })}
       >
         <div className="mono-font">
-          <b>{getInheritedValue()}</b>
+          <b>{inheritedValue}</b>
         </div>
         {inheritedValue && (
           <div className="p-text--small u-text--muted">
@@ -138,7 +108,7 @@ export const getConfigurationRow = ({
       </div>
     ),
     override: isReadOnly ? (
-      getOverrideValue()
+      overrideValue
     ) : isOverridden ? (
       getForm()
     ) : (
