+++ src/pages/projects/forms/DeviceUsageRestrictionForm.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 import { Input, Select } from "@canonical/react-components";
-import { getConfigurationRow } from "pages/instances/forms/ConfigurationRow";
+import { getConfigurationRowInstance } from "pages/instances/forms/ConfigurationRowInstance";
 import ConfigurationTable from "pages/instances/forms/ConfigurationTable";
 import { ProjectFormValues } from "pages/projects/CreateProjectForm";
 import { FormikProps } from "formik/dist/types";
@@ -52,7 +52,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
   return (
     <ConfigurationTable
       rows={[
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_disk",
           label: (
@@ -69,7 +69,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlockManaged} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_disk_paths",
           label: "Disk devices path",
@@ -78,7 +78,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter paths" type="text" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_gpu",
           label: "GPU devices",
@@ -88,7 +88,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_infiniband",
           label: "Infiniband devices",
@@ -98,7 +98,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_nic",
           label: "Network devices",
@@ -109,7 +109,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlockManaged} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_pci",
           label: "PCI devices",
@@ -119,7 +119,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_unix_block",
           label: "Unix-block devices",
@@ -129,7 +129,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_unix_char",
           label: "Unix-char devices",
@@ -139,7 +139,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_unix_hotplug",
           label: "Unix-hotplug devices",
@@ -149,7 +149,7 @@ const DeviceUsageRestrictionForm: FC<Props> = ({ formik }) => {
           children: <Select options={optionAllowBlock} />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "restricted_devices_usb",
           label: "USB devices",
