+++ src/pages/projects/forms/ResourceLimitsForm.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 import { Input } from "@canonical/react-components";
-import { getConfigurationRow } from "pages/instances/forms/ConfigurationRow";
+import { getConfigurationRowInstance } from "pages/instances/forms/ConfigurationRowInstance";
 import ConfigurationTable from "pages/instances/forms/ConfigurationTable";
 import { ProjectFormValues } from "pages/projects/CreateProjectForm";
 import { FormikProps } from "formik/dist/types";
@@ -41,7 +41,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
   return (
     <ConfigurationTable
       rows={[
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_instances",
           label: "Max number of instances",
@@ -50,7 +50,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_containers",
           label: "Max number of containers",
@@ -59,7 +59,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_virtual_machines",
           label: "Max number of VMs",
@@ -68,7 +68,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_disk",
           label: "Max disk space (used by all instances)",
@@ -83,7 +83,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_networks",
           label: "Max number of networks",
@@ -92,7 +92,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_cpu",
           label: "Max sum of individual CPU configurations",
@@ -101,7 +101,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_memory",
           label: "Max sum of individual memory limits",
@@ -110,7 +110,7 @@ const ResourceLimitsForm: FC<Props> = ({ formik }) => {
           children: <Input placeholder="Enter number" min={0} type="number" />,
         }),
 
-        getConfigurationRow({
+        getConfigurationRowInstance({
           formik: formik as unknown as SharedFormikTypes,
           name: "limits_processes",
           label: "Max sum of individual processes configurations",
